# paper/artifacts/phase2/swagger_openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SupplyTwinAI API",
    "description": "Autonomous Multi-Agent Supply Chain Digital Twin REST API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/suppliers/": {
      "get": {
        "tags": [
          "Suppliers"
        ],
        "summary": "List Suppliers",
        "operationId": "list_suppliers_api_v1_suppliers__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierResponse"
                  },
                  "title": "Response List Suppliers Api V1 Suppliers  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/suppliers/{supplier_id}": {
      "get": {
        "tags": [
          "Suppliers"
        ],
        "summary": "Get Supplier",
        "operationId": "get_supplier_api_v1_suppliers__supplier_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Supplier Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SupplierResponse": {
        "properties": {
          "supplier_id": {
            "type": "integer",
            "title": "Supplier Id"
          },
          "global_id": {
            "type": "string",
            "title": "Global Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          },
          "on_time_rate": {
            "type": "number",
            "title": "On Time Rate"
          },
          "defect_rate": {
            "type": "number",
            "title": "Defect Rate"
          },
          "lead_time_days": {
            "type": "integer",
            "title": "Lead Time Days"
          },
          "rating": {
            "type": "number",
            "title": "Rating"
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic"
          }
        },
        "type": "object",
        "required": [
          "supplier_id",
          "global_id",
          "name",
          "category_id",
          "category",
          "country",
          "is_primary",
          "on_time_rate",
          "defect_rate",
          "lead_time_days",
          "rating",
          "is_synthetic"
        ],
        "title": "SupplierResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}